requirements.txt
requests==2.31.0
feedparser==6.0.12
openai==1.12.0
python-dotenv==1.0.0